Move ssvc_calculator to utils.py to avoid code duplication
