+++ vulnerabilities/utils.py
@@ -678,3 +678,118 @@ def create_registry(pipelines):
         registry[key] = pipeline
 
     return registry
+
+
+
+def ssvc_calculator(ssvc_data):
+    """
+    Return the ssvc vector and the decision value
+    """
+    options = ssvc_data.get("options", [])
+    timestamp = ssvc_data.get("timestamp")
+
+    # Extract the options into a dictionary
+    options_dict = {k: v.lower() for option in options for k, v in option.items()}
+
+    # We copied the table value from this link.
+    # https://www.cisa.gov/sites/default/files/publications/cisa-ssvc-guide%20508c.pdf
+
+    # Determining Mission and Well-Being Impact Value
+    mission_well_being_table = {
+        # (Mission Prevalence, Public Well-being Impact) : "Mission & Well-being"
+        ("minimal", "minimal"): "low",
+        ("minimal", "material"): "medium",
+        ("minimal", "irreversible"): "high",
+        ("support", "minimal"): "medium",
+        ("support", "material"): "medium",
+        ("support", "irreversible"): "high",
+        ("essential", "minimal"): "high",
+        ("essential", "material"): "high",
+        ("essential", "irreversible"): "high",
+    }
+
+    if "Mission Prevalence" not in options_dict:
+        options_dict["Mission Prevalence"] = "minimal"
+
+    if "Public Well-being Impact" not in options_dict:
+        options_dict["Public Well-being Impact"] = "material"
+
+    options_dict["Mission & Well-being"] = mission_well_being_table[
+        (options_dict["Mission Prevalence"], options_dict["Public Well-being Impact"])
+    ]
+
+    decision_key = (
+        options_dict.get("Exploitation"),
+        options_dict.get("Automatable"),
+        options_dict.get("Technical Impact"),
+        options_dict.get("Mission & Well-being"),
+    )
+
+    decision_points = {
+        "Exploitation": {"E": {"none": "N", "poc": "P", "active": "A"}},
+        "Automatable": {"A": {"no": "N", "yes": "Y"}},
+        "Technical Impact": {"T": {"partial": "P", "total": "T"}},
+        "Public Well-being Impact": {"B": {"minimal": "M", "material": "A", "irreversible": "I"}},
+        "Mission Prevalence": {"P": {"minimal": "M", "support": "S", "essential": "E"}},
+        "Mission & Well-being": {"M": {"low": "L", "medium": "M", "high": "H"}},
+    }
+
+    # Create the SSVC vector
+    ssvc_vector = "SSVCv2/"
+    for key, value_map in options_dict.items():
+        options_key = decision_points.get(key)
+        for lhs, rhs_map in options_key.items():
+            ssvc_vector += f"{lhs}:{rhs_map.get(value_map)}/"
+
+    # "Decision": {"D": {"Track": "T", "Track*": "R", "Attend": "A", "Act": "C"}},
+    decision_values = {"Track": "T", "Track*": "R", "Attend": "A", "Act": "C"}
+
+    decision_lookup = {
+        ("none", "no", "partial", "low"): "Track",
+        ("none", "no", "partial", "medium"): "Track",
+        ("none", "no", "partial", "high"): "Track",
+        ("none", "no", "total", "low"): "Track",
+        ("none", "no", "total", "medium"): "Track",
+        ("none", "no", "total", "high"): "Track*",
+        ("none", "yes", "partial", "low"): "Track",
+        ("none", "yes", "partial", "medium"): "Track",
+        ("none", "yes", "partial", "high"): "Attend",
+        ("none", "yes", "total", "low"): "Track",
+        ("none", "yes", "total", "medium"): "Track",
+        ("none", "yes", "total", "high"): "Attend",
+        ("poc", "no", "partial", "low"): "Track",
+        ("poc", "no", "partial", "medium"): "Track",
+        ("poc", "no", "partial", "high"): "Track*",
+        ("poc", "no", "total", "low"): "Track",
+        ("poc", "no", "total", "medium"): "Track*",
+        ("poc", "no", "total", "high"): "Attend",
+        ("poc", "yes", "partial", "low"): "Track",
+        ("poc", "yes", "partial", "medium"): "Track",
+        ("poc", "yes", "partial", "high"): "Attend",
+        ("poc", "yes", "total", "low"): "Track",
+        ("poc", "yes", "total", "medium"): "Track*",
+        ("poc", "yes", "total", "high"): "Attend",
+        ("active", "no", "partial", "low"): "Track",
+        ("active", "no", "partial", "medium"): "Track",
+        ("active", "no", "partial", "high"): "Attend",
+        ("active", "no", "total", "low"): "Track",
+        ("active", "no", "total", "medium"): "Attend",
+        ("active", "no", "total", "high"): "Act",
+        ("active", "yes", "partial", "low"): "Attend",
+        ("active", "yes", "partial", "medium"): "Attend",
+        ("active", "yes", "partial", "high"): "Act",
+        ("active", "yes", "total", "low"): "Attend",
+        ("active", "yes", "total", "medium"): "Act",
+        ("active", "yes", "total", "high"): "Act",
+    }
+
+    decision = decision_lookup.get(decision_key, "")
+
+    if decision:
+        ssvc_vector += f"D:{decision_values.get(decision)}/"
+
+    if timestamp:
+        timestamp_formatted = dateparser.parse(timestamp).strftime("%Y-%m-%dT%H:%M:%SZ")
+
+        ssvc_vector += f"{timestamp_formatted}/"
+    return ssvc_vector, decision
